+++ tests/testthat/test-simulate_g2.R
@@ -14,3 +14,28 @@ test_that("g2 results matrix is computed", {
                                        genotypes = msats)$estMat), TRUE)
 })
 
+test_that("g2 results matrix is computed with snps", {
+    expect_equal(is.matrix(simulate_g2(n_ind = 20, subsets = c(2,4), reps = 10, mean_MLH = 0.5, 
+                                       sd_MLH = 0.025, type = "snps")$estMat), TRUE)
+    expect_equal(is.matrix(simulate_g2(n_ind = 20, subsets = c(2,4), type = "snps", reps = 10, 
+                                       genotypes = msats, mean_MLH = 0.5, 
+                                       sd_MLH = 0.025)$estMat), TRUE)
+    expect_equal(is.matrix(simulate_g2(n_ind = 20, subsets = c(2,4), type = "snps", reps = 10, 
+                                       genotypes = msats)$estMat), TRUE)
+})
+
+test_that("subsets have the right range", {
+    expect_error(simulate_g2(n_ind = 20, subsets = 0, reps = 10, mean_MLH = 0.5, 
+                                       sd_MLH = 0.025))
+    expect_error(simulate_g2(n_ind = 20, subsets = c(1,2), reps = 10, mean_MLH = 0.5, 
+                             sd_MLH = 0.025))
+    expect_error(simulate_g2(n_ind = 20, subsets = NULL, reps = 10, mean_MLH = 0.5, 
+                             sd_MLH = 0.025))
+})
+
+test_that("MLH have the right range", {
+    expect_error(simulate_g2(n_ind = 20, subsets = 0, reps = 10, mean_MLH = 1.5, 
+                             sd_MLH = 0.025))
+    expect_error(simulate_g2(n_ind = 20, subsets = c(1,2), reps = 10, mean_MLH = -0.5, 
+                             sd_MLH = 0.025))
+})